S8/smart-agent/requirements.txt
# requirements.txt
openai==1.55.3
gspread
oauth2client
pydantic
flask
fastapi
uvicorn
python-dotenv==1.0.1
mcp
aiosmtplib
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-telegram-bot==20.0
uvicorn[standard]==0.27.1

